requirements.txt
biopython
numpy
pandas
scikit-learn
pytest
python_codon_tables
weblogo
